Separated database package
