+++ app/src/main/java/com/example/runtracker/RunApplication.kt
@@ -1,6 +1,8 @@
 package com.example.runtracker
 
 import android.app.Application
+import com.example.runtracker.database.RunDatabase
+import com.example.runtracker.database.RunRepository
 
 class RunApplication: Application() {
     private val database by lazy { RunDatabase.getDatabase(this) }
