+++ app/src/main/java/com/example/runtracker/SummaryActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.widget.TextView
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import com.example.runtracker.database.Run
 import kotlin.math.floor
 import kotlin.math.round
