+++ app/src/main/java/com/example/runtracker/database/Run.kt
@@ -1,4 +1,4 @@
-package com.example.runtracker
+package com.example.runtracker.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
